requirements.txt
torch
torchvision
easydict==1.13
einops==0.8.0
imageio==2.34.2
keras==3.3.3
ml-dtypes==0.3.2
numpy==1.24.4
openpyxl==3.1.5
pandas==2.2.2
pillow==10.3.0
pyyaml==6.0.1
requests==2.31.0
scikit-image==0.24.0
scikit-learn==1.1.3
scipy==1.13.0
setuptools==65.5.0
six==1.16.0
sympy==1.11.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.7.24
tqdm==4.66.2